requirements.txt
Faker==13.12.1
python-dateutil==2.8.2
pyzmq==23.1.0
six==1.16.0
